final project for MATH 3333
Final project: 20% of overall grade.

reference paper section 5.1 for the description of the data
photoMetaData (downloaded from github)
columbiaImages (downloaded from github)
math3333projectsamplecode (acknowledgement:from online source)
Project description: 1) Read the reference paper to know the background information about the collection of the 800 Columbia images. 2) Use the sample code provided above to perform the basic glm analysis to predict the outcome. Here the outcome is whether or not the image is an outdoor picture. (note: change the file path name to your own, do not use your teacher’s computer path name.) 3) The predictors in the sample code are the 3 medians of the pixel intensities of the images. You need to develop a new method to extract the information from the image. A suggestion is to divide image into grids and get the median pixels of different gird areas. You may use other ways to summarize the pixel intensities. You are encouraged to use all the pixel information and then use variable selection method such as lasso to build a concise model. You are encouraged to use other methods we learned in this course to develop or modify the basic method provided in the sample code. 4) Write a formal report of 10 pages with the following sections: title, abstract, introduction, data, methodology, analysis and model, result, conclusion, references. This report is individual work and no plagiarism is allowed. Due date: last day of class, April 2, 2020. The marked paper will be returned during the final exam.